+++ Assets/WorldGeneration/_Scripts/WorldManager.cs
@@ -10,27 +10,29 @@ namespace WorldGeneration._Scripts
     public class WorldManager : MonoBehaviour
     {
         // [SerializeField]
-        [Header("General Settings")]
-        [SerializeField] private Vector2Int resolution = new(128, 128);
+        [Header("General Settings")] [SerializeField]
+        private Vector2Int resolution = new(128, 128);
+
         [SerializeField] private float maxTerrainHeight = 180.0f;
-        [Range(0.0f, 1.0f)][SerializeField] private float waterLevel = 0.3f;
+        [Range(0.0f, 1.0f)] [SerializeField] private float waterLevel = 0.3f;
         [SerializeField] private GeneralSettings generalSettings;
         [SerializeField] private NoiseSettings noiseSettings;
-        
-        [Header("Objects")]
-        [SerializeField] private GameObject wallPrefab;
+
+        [Header("Objects")] [SerializeField] private GameObject wallPrefab;
+
         [SerializeField] private GameObject demoCarrot;
         [SerializeField] private int maxNumberOfCarrots = 20;
+        private AssetManager _assetManager;
+        private NoiseWithClamp _noiseWithClamp;
+
+        private float[,] _map;
+        private Vector2[] _corners;
+
+        private bool _running;
 
         // private
         private TerrainManager _terrainManager;
-        private AssetManager _assetManager;
         private TimeManager _timer;
-
-        private NoiseWithClamp _noiseWithClamp;
-
-        private bool _running;
-        private Vector2[] _corners;
 
         // Start is called before the first frame update
         private void Start()
@@ -39,16 +41,16 @@ namespace WorldGeneration._Scripts
             _assetManager = AssetManager.GetInstance();
             _timer = TimeManager.Instance;
 
-            _noiseWithClamp.NoiseGenerator = new NoiseGenerator();
+            _noiseWithClamp.NoiseGenerator = new NoiseGenerator(noiseSettings);
             _noiseWithClamp.ValueClamp = new ValueClamp();
 
-            _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, wallPrefab,
-                _noiseWithClamp);
+            _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings,
+                wallPrefab, _noiseWithClamp, out _map);
 
-            _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
+            // _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
 
-            _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _noiseWithClamp,
-                demoCarrot, _corners, maxNumberOfCarrots);
+            _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _map,
+                _noiseWithClamp, demoCarrot, _corners, maxNumberOfCarrots);
 
             _running = true;
         }
@@ -77,13 +79,13 @@ namespace WorldGeneration._Scripts
             // Reload Terrain
             if (Input.GetKey(KeyCode.R) && _running)
             {
-                _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, wallPrefab,
-                    _noiseWithClamp);
+                _terrainManager.GenerateTerrain(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings,
+                    wallPrefab, _noiseWithClamp, out _map);
 
-                _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
+                // _corners = GeneratorFunctions.GetCornerPoints(generalSettings, noiseSettings, _noiseWithClamp);
 
-                _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _noiseWithClamp,
-                    demoCarrot, _corners, maxNumberOfCarrots);
+                _assetManager.SpawnAssets(resolution, maxTerrainHeight, waterLevel, generalSettings, noiseSettings, _map,
+                    _noiseWithClamp, demoCarrot, _corners, maxNumberOfCarrots);
             }
         }
     }