+++ Changelog.md
@@ -1,6 +1,14 @@
 ﻿# Changelog World Generation
 
 
+## [0.7.0] - 2023-04-13
+# Added
+- Random Spawner for carrots
+
+# Changed
+- OpenSimplexNoise instead of Perlin Noise
+
+
 ## [0.6.0] - 2023-04-12
 - Code Refactoring